# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenMuara
  description: Local-first billing and payment virtualization layer.
  version: 0.1.1
  contact:
    name: OpenMuara
servers:
  - url: http://127.0.0.1:9000
    description: Local development server
paths:
...
  /v1/pay:
    post:
      summary: Create a provider-agnostic payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '200':
          description: Existing payment returned (idempotency hit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/pay/{ref}:
    get:
      summary: Get a payment by reference
      operationId: getPayment
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/refund/{ref}:
    post:
      summary: Refund a payment by reference
      operationId: refundPayment
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Refund recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /_admin/stripe/webhooks:
    get:
      summary: Stripe webhook configuration page
      operationId: stripeWebhookConfigPage
      responses:
        '200':
          description: HTML webhook config page
          content:
            text/html:
              schema:
                type: string
    post:
      summary: Save Stripe webhook configuration
      operationId: stripeWebhookConfigSave
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                webhook_url:
                  type: string
                enabled_events:
                  type: array
                  items:
                    type: string
                csrf_token:
                  type: string
      responses:
        '302':
          description: Redirects back to config page
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — CSRF token missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict — invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            trace_id:
              type: string
    TransactionItem:
      type: object
      properties:
        itemCode:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
    Transaction:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        status:
          type: string
          enum: [new, paid, unpaid, refunded]
        customerRef:
          type: string
        idempotencyKey:
          type: string
        reference:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentRequest:
      type: object
      required: [provider, type, amount, currency, reference]
      properties:
        provider:
          type: string
        type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        customer_ref:
          type: string
        reference:
          type: string
        idempotency_key:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
    PaymentResponse:
      allOf:
        - $ref: '#/components/schemas/Transaction'
...